Big change - client db versioning. The former qs-init event is now qs-load. And this is where plugins should register the services they provide, but they are not allowed to interact with the database. All database interaction should be performed in the *new* qs-init.
The reason for this is the changes needed in order to support client database versioning. And one of the things that plugins register in qs-load is themselves, complete with version number. And that version number is checked between load and init. The database schema and other things can therefore not be guaranteed to be correct at qs-load.

A version number that doesn't match blocks normal quicksilver operation and shows a "Update needed"-message with a link to the update page when the normal hotkeys are pressed.

A dummy update page is in place, but no update scripts are loaded.